scripts: enhance cray runall script
Add support for running the client/server tests
in the cray test script.  Currently they are run
using the sockets provider.  We can eventually
add a subset list of the client/server tests that
will be runnable using the gni provider.

Currently some of the tests segfault if the
-f sockets
argument is added to the test command line.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>